+++ src/MLJ.jl
@@ -22,7 +22,7 @@ export ConstantRegressor, ConstantClassifier         # "builtins/Constant.jl
 export DeterministicConstantRegressor                # "builtins/Constant.jl
 export DeterministicConstantClassifier               # "builtins/Constant.jl
 export KNNRegressor                                  # "builtins/KNN.jl"
-export RidgeRegressor                                # "builtins/LocalMulitivariateStats.jl
+export RidgeRegressor, PCA                           # "builtins/LocalMulitivariateStats.jl
 
 # defined in include files "machines.jl" and "networks.jl":
 export Machine, NodalMachine, machine
@@ -132,7 +132,6 @@ function __init__()
     @load_interface DecisionTree "7806a523-6efd-50cb-b5f6-3fa6f1930dbb" lazy=true
     @load_interface GaussianProcesses "891a1506-143c-57d2-908e-e1f8e92e6de9" lazy=true
     @load_interface Clustering "aaaa29a8-35af-508c-8bc3-b662a17a0fe5" lazy=true
-    @load_interface MultivariateStats "6f286f6a-111f-5878-ab1e-185364afe411" lazy=true
 #    @load_interface  MultivariateStats "6f286f6a-111f-5878-ab1e-185364afe411" lazy=true
 end
