+++ src/builtins/LocalMultivariateStats.jl
@@ -2,12 +2,14 @@
 
 module LocalMultivariateStats
 
-export RidgeRegressor 
+export RidgeRegressor, PCA
 
 import MLJBase
 import MLJ
 import MultivariateStats
 import DataFrames
+
+const MS = MultivariateStats
 
 struct LinearFitresult <: MLJBase.MLJType
     coefficients::Vector{Float64}
@@ -42,6 +44,10 @@ function coef_info(fitresult::LinearFitresult, features)
     return df
 end
 
+####
+#### RIDGE
+####
+
 mutable struct RidgeRegressor <: MLJBase.Deterministic{LinearFitresult}
     lambda::Float64
 end
@@ -59,7 +65,7 @@ function MLJBase.fit(model::RidgeRegressor, verbosity::Int, Xplus, y::Vector{<:R
 
     X, features = Xplus
 
-    weights = MultivariateStats.ridge(X, y, model.lambda)
+    weights = MS.ridge(X, y, model.lambda)
 
     coefficients = weights[1:end-1]
     bias = weights[end]
@@ -106,6 +112,92 @@ MLJBase.inputs_can_be(::Type{<:RidgeRegressor}) = [:numeric, ]
 MLJBase.target_kind(::Type{<:RidgeRegressor}) = :numeric
 MLJBase.target_quantity(::Type{<:RidgeRegressor}) = :univariate
 
+####
+#### PCA
+####
+
+const PCAFitResultType = MS.PCA
+
+mutable struct PCA <: MLJBase.Unsupervised
+    ncomp::Union{Nothing, Int} # number of PCA components, all if nothing
+    method::Symbol  # cov or svd (auto by default, choice based on dims)
+    pratio::Float64 # ratio of variances preserved in the principal subspace
+    mean::Union{Nothing, Real, Vector{Float64}} # 0 if pre-centered
+end
+
+function PCA(; ncomp=nothing
+             , method=:auto
+             , pratio=0.99
+             , mean=nothing)
+
+    model = PCA(ncomp, method, pratio, mean)
+    message = MLJBase.clean!(model)
+    isempty(message) || @warn message
+    return model
+end
+
+function MLJBase.clean!(model::PCA)
+    warning = ""
+    if model.ncomp isa Int && model.ncomp < 1
+        warning *= "Need ncomp > 1. Resetting ncomp=p.\n"
+        model.ncomp = nothing
+    end
+    if model.method ∉ [:auto, :cov, :svd]
+        warning *= "Unknown method specification. Resetting to method=:auto.\n"
+        model.method = :auto
+    end
+    if !(0.0 < model.pratio <= 1.0)
+        warning *= "Need 0 < pratio < 1. Resetting to pratio=0.99.\n"
+        model.pratio = 0.99
+    end
+    if (model.mean isa Real) && !iszero(model.mean)
+        warning *= "Need mean to be nothing, zero or a vector." *
+                   " Resetting to mean=nothing.\n"
+        model.mean = nothing
+    end
+    return warning
+end
+
+function MLJBase.fit(model::PCA
+                   , verbosity::Int
+                   , X)
+
+    Xarray = MLJBase.matrix(X)
+    mindim = minimum(size(Xarray))
+
+    ncomp = (model.ncomp === nothing) ? mindim : model.ncomp
+
+    # NOTE: copy/transpose
+    fitresult = MS.fit(MS.PCA, copy(transpose(Xarray))
+                     ; method=model.method
+                     , pratio=model.pratio
+                     , maxoutdim=ncomp
+                     , mean=model.mean)
+
+    cache = nothing
+    report = nothing
+
+    return fitresult, cache, report
+end
+
+function MLJBase.transform(model::PCA
+                         , fitresult::PCAFitResultType
+                         , X)
+
+    Xarray = MLJBase.matrix(X)
+    # X is n x d, need to transpose and copy twice...
+    return copy(transpose(MS.transform(fitresult, copy(transpose(Xarray)))))
+end
+
+####
+#### METADATA
+####
+
+MLJBase.package_name(::Type{PCA})  = "MultivariateStats"
+MLJBase.package_uuid(::Type{PCA})  = "6f286f6a-111f-5878-ab1e-185364afe411"
+MLJBase.is_pure_julia(::Type{PCA}) = :yes
+MLJBase.inputs_can_be(::Type{PCA}) = [:numeric,]
+
 end # of module
 
 
@@ -115,4 +207,3 @@ using .LocalMultivariateStats
 
 #using .MultivariateStats_
 #export RidgeRegressor
-
